# docs/swagger.yaml
basePath: /
definitions:
...
  hotel-management_internal_dto.MailRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  hotel-management_internal_dto.SearchRoomRequest:
    properties:
      bed_num:
        type: integer
      end_date:
        type: string
      has_aircon:
        type: boolean
      max_price:
        type: number
      min_price:
        type: number
      start_date:
        type: string
      view_type:
        type: string
    required:
    - end_date
    - start_date
    type: object
  hotel-management_internal_dto.SearchRoomResponse:
    properties:
      bed_num:
        type: integer
      description:
        type: string
      has_aircon:
        type: boolean
      id:
        type: integer
      image_urls:
        items:
          type: string
        type: array
      name:
        type: string
      price_per_night:
        type: number
      type:
        type: string
      view_type:
        type: string
    type: object
  hotel-management_internal_dto.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      phone_number:
        type: string
    required:
    - name
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is an API for hotel management.
  title: Hotel Management API
  version: "1.0"
paths:
...
  /mail/reset-password:
    post:
      consumes:
      - application/json
      description: Generates a new password and sends it via email to the user.
      parameters:
      - description: User email for password reset
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hotel-management_internal_dto.MailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: New password sent.'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'message: Invalid request data.'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'message: Failed to get user'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Failed to update user.'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset user password
      tags:
      - Mail
  /mail/smtp-verify:
    post:
      consumes:
      - application/json
      description: Send a verification email with token to user's email address if
        the account is not activated.
      parameters:
      - description: Email to send verification link
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hotel-management_internal_dto.MailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verification email sent.
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request data. | User already verified.
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Failed to get user.
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Could not generate token. | Failed to send verification email.
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send verification email
      tags:
      - Mail
  /mail/verify-account:
    get:
      consumes:
      - application/json
      description: Activate user account based on the token sent via email
      parameters:
      - description: JWT token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account verified successfully.
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid token. | User already verified.
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Failed to get user.
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update user.
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Activate user account
      tags:
      - Mail
  /payments/vnpay_return:
    get:
      description: Handle the response returned from VnPay after a payment attempt.
      parameters:
      - description: Transaction Reference
        in: query
        name: vnp_TxnRef
        required: true
        type: string
      - description: VnPay Response Code
        in: query
        name: vnp_ResponseCode
        required: true
        type: string
      - description: VnPay Transaction Number
        in: query
        name: vnp_TransactionNo
        required: true
        type: string
      responses:
        "200":
          description: Payment processed successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid callback parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Payment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Payment has already been processed or booking already paid
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update payment, process payment, or create bill
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Process VnPay callback
      tags:
      - payments
...
  /rooms/search:
    post:
      consumes:
      - application/json
      description: Find all available rooms that match the search criteria and are
        not booked during the requested time range
      parameters:
      - description: Search filters for room availability
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hotel-management_internal_dto.SearchRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Find available room successful!
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/hotel-management_internal_dto.SearchRoomResponse'
              type: array
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to find available room.
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search available rooms
      tags:
      - Rooms
  /users/update-profile:
    put:
      consumes:
      - application/json
      description: Update the name, avatar, or phone number of the current user
      parameters:
      - description: Update profile request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/hotel-management_internal_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Update profile successful.'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'message: Invalid request data.'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Failed to get user'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Failed to update user. or Internal server error.'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"